Refactor date handling to standardize `toIsoDate` utility usage and migrate `occurredAt` and `startDate` fields to ISO string format across components and database operations.
