+++ lib/utils.ts
@@ -98,6 +98,98 @@ export function calculateNextRenewal(
   return nextRenewal
 }
 
+/**
+ * Calculates all past renewal dates between a start date and the current date
+ * @param startDate The start date of the subscription
+ * @param billingCycle The billing cycle of the subscription
+ * @param currentDate The current date (defaults to now)
+ * @returns An array of dates representing past renewals (excluding the initial date)
+ */
+export function calculatePastRenewals(
+  startDate: Date,
+  billingCycle: string,
+  currentDate: Date = new Date()
+): Date[] {
+  // Create copies of dates to avoid modifying the originals
+  const start = new Date(startDate);
+  const current = new Date(currentDate);
+
+  // Set the time to midnight to avoid time comparison issues
+  start.setHours(0, 0, 0, 0);
+  current.setHours(0, 0, 0, 0);
+
+  // If start date is in the future or today, there are no past renewals
+  if (start >= current) {
+    return [];
+  }
+
+  const renewalDates: Date[] = [];
+  const nextRenewal = new Date(start);
+
+  // Calculate all renewal dates until we reach or exceed the current date
+  switch (billingCycle) {
+    case "daily":
+      while (nextRenewal < current) {
+        nextRenewal.setDate(nextRenewal.getDate() + 1);
+        if (nextRenewal < current) {
+          renewalDates.push(new Date(nextRenewal));
+        }
+      }
+      break;
+
+    case "weekly":
+      while (nextRenewal < current) {
+        nextRenewal.setDate(nextRenewal.getDate() + 7);
+        if (nextRenewal < current) {
+          renewalDates.push(new Date(nextRenewal));
+        }
+      }
+      break;
+
+    case "monthly":
+      while (nextRenewal < current) {
+        nextRenewal.setMonth(nextRenewal.getMonth() + 1);
+        if (nextRenewal < current) {
+          renewalDates.push(new Date(nextRenewal));
+        }
+      }
+      break;
+
+    case "quarterly":
+      while (nextRenewal < current) {
+        nextRenewal.setMonth(nextRenewal.getMonth() + 3);
+        if (nextRenewal < current) {
+          renewalDates.push(new Date(nextRenewal));
+        }
+      }
+      break;
+
+    case "yearly":
+      while (nextRenewal < current) {
+        nextRenewal.setFullYear(nextRenewal.getFullYear() + 1);
+        if (nextRenewal < current) {
+          renewalDates.push(new Date(nextRenewal));
+        }
+      }
+      break;
+
+    case "one_time":
+      // No renewals for one-time subscriptions
+      break;
+
+    default:
+      // Default to monthly if the billing cycle is not recognized
+      while (nextRenewal < current) {
+        nextRenewal.setMonth(nextRenewal.getMonth() + 1);
+        if (nextRenewal < current) {
+          renewalDates.push(new Date(nextRenewal));
+        }
+      }
+  }
+
+  return renewalDates;
+}
+
 export function toIsoDate(date: Date): string {
   return format(date, "yyyy-MM-dd");
 }
